refactor: mv provider to provider/signing
