+++ src/provider/signing.test.ts
@@ -1,6 +1,7 @@
-import { JsonRpcProvider, JsonRpcSigner } from '@ethersproject/providers'
+import { ExternalProvider, JsonRpcProvider, JsonRpcSigner, Web3Provider } from '@ethersproject/providers'
+import { Mutable } from 'types'
 
-import { signTypedData } from './provider'
+import { signTypedData } from '../provider'
 
 describe('provider', () => {
   describe('signTypedData', () => {
@@ -103,7 +104,8 @@ describe('provider', () => {
     describe('wallets which do not support eth_signTypedData_v4', () => {
       describe.each(['SafePal Wallet'])('%s', (name) => {
         beforeEach(() => {
-          signer.provider['provider'] = { isWalletConnect: true, connector: { peerMeta: { name } } }
+          const web3Provider = signer.provider as Mutable<Web3Provider>
+          web3Provider.provider = { isWalletConnect: true, connector: { peerMeta: { name } } } as ExternalProvider
         })
 
         it('signs using eth_signTypedData', async () => {
